@@ -1,6 +1,6 @@
 module github.com/ti-mo/conntrack
 
-go 1.12
+go 1.18
 
 require (
 	github.com/google/go-cmp v0.5.9
@@ -10,4 +10,14 @@ require (
 	github.com/ti-mo/netfilter v0.5.0
 	github.com/vishvananda/netns v0.0.0-20180720170159-13995c7128cc
 	golang.org/x/sys v0.2.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/josharian/native v1.0.0 // indirect
+	github.com/mdlayher/socket v0.4.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/net v0.2.0 // indirect
+	golang.org/x/sync v0.1.0 // indirect
+	gopkg.in/yaml.v2 v2.2.2 // indirect
 )
